+++ src/constants/form-element.ts
@@ -10,16 +10,17 @@ export const enum FormElementType {
   CODE = 'code',
   DATETIME = 'datetime',
   DISABLED = 'disabled',
+  DISABLED_TEXTAREA = 'disabledTextarea',
+  FILE = 'file',
+  MULTISELECT = 'multiselect',
   NUMBER = 'number',
   PASSWORD = 'password',
   RADIO = 'radio',
   SECRET = 'secret',
   SELECT = 'select',
-  MULTISELECT = 'multiselect',
   SLIDER = 'slider',
   STRING = 'string',
   TEXTAREA = 'textarea',
-  FILE = 'file',
 }
 
 /**
@@ -65,6 +66,10 @@ export const FormElementTypeOptions: SelectableValue[] = [
   {
     value: FormElementType.DISABLED,
     label: 'Read-only',
+  },
+  {
+    value: FormElementType.DISABLED_TEXTAREA,
+    label: 'Read-only Text Area',
   },
   {
     value: FormElementType.SELECT,